+++ sites/docs/CHANGELOG.md
@@ -1,5 +1,29 @@
 # evidence-docs
 
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies [6ecef713f]
+- Updated dependencies [7a414e88e]
+- Updated dependencies [ddf5da4d0]
+- Updated dependencies [c82ba1516]
+- Updated dependencies [320c39941]
+- Updated dependencies [8778bcfc3]
+- Updated dependencies [011a08be7]
+- Updated dependencies [475c42432]
+- Updated dependencies [32690c270]
+- Updated dependencies [7f70975e5]
+- Updated dependencies [c17626b1c]
+- Updated dependencies [667b7167e]
+- Updated dependencies [58403f1db]
+- Updated dependencies [4e240652b]
+  - @evidence-dev/core-components@4.8.11
+  - @evidence-dev/sdk@1.4.6
+  - @evidence-dev/evidence@39.1.14
+  - @evidence-dev/component-utilities@3.2.18
+  - @evidence-dev/tailwind@2.1.7
+
 ## 0.0.39
 
 ### Patch Changes